feat(mock): Mock controller 에서 테스트 데이터 관리
